+++ Source/ACE.DatLoader/DatFileType.cs
@@ -324,7 +324,7 @@ namespace ACE.DatLoader
         MaterialInstance        = 33, // DB_TYPE_MATERIALINSTANCE
 
         /// <summary>
-        /// suspected to be Submodel Table.  rename this if it can be confirmed.
+        /// SoundTable
         /// </summary>
         [DatDatabaseType(DatDatabaseType.Portal)]
         [DatFileTypeExtension("stb")]
@@ -332,8 +332,9 @@ namespace ACE.DatLoader
         STable                  = 34, // DB_TYPE_STABLE
 
         /// <summary>
-        /// client pdb implies this was in a server-only file.
+        /// This is in the Language dat (client_local_English.dat)
         /// </summary>
+        [DatDatabaseType(DatDatabaseType.Portal)]
         [DatFileTypeExtension("uil")]
         [DatFileTypeIdRange(0x21000000, 0x21FFFFFF)]
         UiLayout                = 35, // DB_TYPE_UI_LAYOUT
@@ -347,8 +348,9 @@ namespace ACE.DatLoader
         EnumMapper              = 36, // DB_TYPE_ENUM_MAPPER
 
         /// <summary>
-        /// client pdb implies this was in a server-only file.
+        /// This is in the Language dat (client_local_English.dat)
         /// </summary>
+        [DatDatabaseType(DatDatabaseType.Portal)]
         [DatFileTypeExtension("stt")]
         [DatFileTypeIdRange(0x23000000, 0x24FFFFFF)]
         StringTable             = 37, // DB_TYPE_STRING_TABLE 
@@ -422,9 +424,10 @@ namespace ACE.DatLoader
         FontLocal               = 47, // DB_TYPE_FONT_LOCAL 
 
         /// <summary>
-        /// client pdb implies this was in a server-only file.  also, indexed as
+        /// This is located in the Language dat (client_local_English.dat)
         /// "stringtable" in the client
         /// </summary>
+        [DatDatabaseType(DatDatabaseType.Portal)]
         [DatFileTypeExtension("sti")]
         [DatFileTypeIdRange(0x41000000, 0x41FFFFFF)]
         StringState = 48, // DB_TYPE_STRING_STATE